U1-99 Update intefration test afterEach to avoid exception

+++ tests/integration/get-images.test.js
@@ -14,7 +14,9 @@ describe('Feature get-images', () => {
   });
 
   afterEach(() => {
-    fs.unlinkSync(path.join(cwd, '\\.userconfig'));
+    if (fs.existsSync(path.join(cwd, '\\.userconfig'))) {
+      fs.unlinkSync(path.join(cwd, '\\.userconfig'));
+    }
   });
 
   it('Should get images with default settings', (done) => {
